refactor (mountain): removing use statements from model module
