+++ mountain/src/main.rs
@@ -28,7 +28,9 @@ use crate::handlers::{add_skier, piste_builder};
 use crate::handlers::{lift_builder, selection};
 use crate::init::generate_heightmap;
 use crate::model::frame::Frame;
-use crate::model::{skiing, Lift, Piste, PisteCosts};
+use crate::model::lift::Lift;
+use crate::model::piste::{Piste, PisteCosts};
+use crate::model::skiing;
 use crate::services::id_allocator;
 use crate::systems::{avatar_artist, cost_computer, framer, lift, lift_entry, overlay, planner};
