+++ mountain/src/model/mod.rs
@@ -1,10 +1,5 @@
-mod direction;
+pub mod direction;
 pub mod frame;
-mod lift;
-mod piste;
+pub mod lift;
+pub mod piste;
 pub mod skiing;
-
-pub use direction::*;
-pub use lift::*;
-pub use piste::*;
-pub use skiing::*;
