+++ mountain/src/systems/cost_computer.rs
@@ -3,8 +3,10 @@ use std::collections::{HashMap, HashSet};
 use commons::geometry::XY;
 use commons::grid::Grid;
 
+use crate::model::direction::DIRECTIONS;
+use crate::model::lift::Lift;
+use crate::model::piste::{Piste, PisteCosts};
 use crate::model::skiing::{Mode, State};
-use crate::model::{Lift, Piste, PisteCosts, DIRECTIONS};
 use crate::network::skiing::{SkiingInNetwork, SkiingNetwork};
 use crate::network::velocity_encoding::{encode_velocity, VELOCITY_LEVELS};
 use network::algorithms::costs_to_target::CostsToTarget;
